refactor v2 transform tests
